AutoFormFiller/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
anthropic>=0.40.0
openai>=1.0.0
google-generativeai>=0.8.0
httpx<0.28.0
python-dotenv==1.0.0
pypdf>=4.0.0
requests>=2.31.0
gunicorn>=21.2.0
